feat: update mz to have result handling as a parameter (#50)
Updates the mz to have the result as a parameter, too.

After some discussion I came to the conclusion that it is best to do it
in the same way as the qubits.

This means that the pointer to the qubits and the result will not match
and this is what is needed here.

---------

Co-authored-by: Travis Thompson <102229498+trvto@users.noreply.github.com>